requirements.txt
numpy==1.9.2
path.py==7.3
scipy==0.15.1
theanify==0.0.1
Theano==0.7.0
